requirements.txt
flask
requests
python-dotenv
markdown2